fix: extracting context
