+++ src/app/transactions/upload/steps/categorize-columns.tsx
@@ -7,7 +7,7 @@ import { useToast } from "@/app/components/ui/use-toast";
 import { addRowToMatrix } from "../../components/operators";
 import { Button } from "@/app/components/ui/button";
 import { TransactionsTable } from "../../components/transactions-table";
-import { UploadTransactionsContext } from "../page";
+import { UploadTransactionsContext } from "@/lib/context";
 
 const headerOptions = ["date", "concept", "amount"];
