+++ src/lib/context.ts
@@ -1,0 +1,18 @@
+import { createContext } from "react";
+
+interface TransactionsContext {
+  transactions: string[][];
+  setTransactions: (transactions: string[][]) => void;
+  nextStep: () => void;
+  uploadTransactions: (matrix: string[][]) => void;
+}
+
+const defaultContext: TransactionsContext = {
+  transactions: [],
+  setTransactions: () => {},
+  nextStep: () => {},
+  uploadTransactions: () => {},
+};
+
+export const UploadTransactionsContext =
+  createContext<TransactionsContext>(defaultContext);
